% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalgamation.R
\name{amalgamation}
\alias{amalgamation}
\title{Amalgamation of levels}
\usage{
amalgamation(bnfit, node)
}
\arguments{
\item{bnfit}{object of class \code{bn.fit}.}

\item{node}{a node of \code{bnfit}}
}
\value{
A list where each entry refers to a child of \code{node}. For each child, the function reports the diameter resulting from the amalgamation of any pair of levels.
}
\description{
Computation of the diameter of children conditional probability tables when levels of a variable are merged.
}
\examples{
amalgamation(synthetic_bn, "y1")


}
\references{
Leonelli, M., Smith, J. Q., & Wright, S. K. (2024). The diameter of a stochastic matrix: A new measure for sensitivity analysis in Bayesian networks. arXiv preprint arXiv:2407.04667.
}
