% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachexia.R
\docType{data}
\name{cachexia}
\alias{cachexia}
\alias{cachexia_gbn}
\alias{cachexia_ci}
\alias{control_gbn}
\alias{control_ci}
\title{Bayesian networks for a cachexia study}
\format{The Bayesian network \code{bnsens_example} comprehends the following nodes:
\itemize{
  \item \bold{y1}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y2}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y3}:  three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
}}
\source{
Manuele Leonelli, Eva Riccomagno (2018). "A geometric characterisation of sensitivity analysis in monomial models". \url{https://arxiv.org/abs/1901.02058}
}
\usage{
cachexia_gbn

cachexia_ci

control_gbn

control_ci
}
\description{
Some comments here
}
\seealso{
Other datasets: 
\code{\link{fire_alarm}},
\code{\link{synt_gbn}},
\code{\link{synthetic_bn}},
\code{\link{travel}}
}
\concept{datasets}
\keyword{datasets}
