% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{covariance_var}
\alias{covariance_var}
\title{Standard variation of the covariance matrix}
\usage{
covariance_var(gbn, entry, delta)
}
\arguments{
\item{gbn}{object of class \code{GBN}.}

\item{entry}{a vector of length 2 specifying the entry of the covariance matrix to vary.}

\item{delta}{additive variation coefficient for the entry of the mean vector given in \code{entry}.}
}
\description{
Computation of an updated \code{GBN} object after a variation of the covariance matrix.
}
\details{
Need to write this description, write examples and explain the return
}
\references{
Gómez-Villegas, M. A., Maín, P., & Susi, R. (2007). Sensitivity analysis in Gaussian Bayesian networks using a divergence measure. Communications in Statistics—Theory and Methods, 36(3), 523-539.

Gómez-Villegas, M. A., Main, P., & Susi, R. (2013). The effect of block parameter perturbations in Gaussian Bayesian networks: Sensitivity and robustness. Information Sciences, 222, 439-458.

 @importFrom matrixcalc is.positive.semi.definite
}
\seealso{
\code{\link{mean_var}}, \code{\link{model_pres_cov}}
}
