% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{Fro.GBN}
\alias{Fro.GBN}
\title{Frobenius norm for \code{GBN}}
\usage{
\method{Fro}{GBN}(x, entry, delta, ...)
}
\arguments{
\item{x}{object of class \code{GBN}.}

\item{entry}{a vector of length 2 indicating the entry of the covariance matrix to vary.}

\item{delta}{variation parameter that acts additively.}

\item{...}{additional parameters to be added to the plot.}
}
\description{
\code{Fro.GBN} returns the Frobenius norm between a Gaussian Bayesian network and its update after an additive parameter variation.
}
