% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{Fro}
\alias{Fro}
\title{Frobenius norm}
\usage{
Fro(x, ...)
}
\arguments{
\item{x}{object of class \code{GBN} or \code{CI}.}

\item{...}{parameters specific to the class used.}
}
\value{
A dataframe whose columns depend of the class of the object.
}
\description{
\code{Fro} returns the Frobenius norm between a Bayesian network and its update after parameter variation.
}
\details{
The details depend on the class the method \code{Fro} is applied to.
}
\seealso{
\code{\link{KL.GBN}}, \code{\link{KL.CI}}, \code{\link{Fro.CI}}, \code{\link{Fro.GBN}}, \code{\link{Jeffreys.GBN}}, \code{\link{Jeffreys.CI}}
}
