% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_subsidiary.R
\name{get.pa.combo.score}
\alias{get.pa.combo.score}
\title{subsidiary functions for global.monitor.bn.node 
#'}
\usage{
get.pa.combo.score(k, counts.vec, alpha.vec)
}
\arguments{
\item{k}{}

\item{alpha.vec}{}
}
\description{
subsidiary functions for global.monitor.bn.node 
#'
}
