% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_monitor.R
\name{global.monitor.bn.node}
\alias{global.monitor.bn.node}
\alias{global.monitor.tbl}
\alias{global.monitor.graph}
\title{global monitors and the dependent functions}
\usage{
global.monitor.bn.node(node.name, dag, alpha, df)

global.monitor.tbl(dag, alpha, df)

global.monitor.graph(dag, alpha, df)
}
\arguments{
\item{node.name}{name from the dag nodes}

\item{dag}{bnlearn object specifying a dag}

\item{alpha}{single integer, usually the number of max levels in df}

\item{df}{a base R style datafrmae}
}
\description{
global monitors and the dependent functions
}
\section{Functions}{
\itemize{
\item \code{global.monitor.tbl}: 

\item \code{global.monitor.graph}: 
}}

