% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_monitor.R
\name{global}
\alias{global}
\alias{global.monitor.bn.node}
\alias{global.monitor}
\alias{global.monitor.graph}
\title{Global monitors and the dependent functions}
\usage{
global.monitor.bn.node(node.idx, dag, alpha, df)

global.monitor(dag, alpha, df)

global.monitor.graph(dag, alpha, df)
}
\arguments{
\item{node.idx}{index from the dag nodes}

\item{dag}{bnlearn object specifying a dag}

\item{alpha}{single integer, usually the number of max levels in df}

\item{df}{a base R style dataframe}
}
\description{
Global monitors and the dependent functions
}
