% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{KL}
\alias{KL}
\title{KL Divergence}
\usage{
KL(x, ...)
}
\arguments{
\item{x}{object of class \code{bn.fit}, \code{GBN} or \code{CI}.}

\item{...}{parameters specific to the class used.}
}
\description{
\code{KL} returns the Kullback-Leibler (KL) divergence between a Bayesian network and its update after parameter variation.
}
\details{
The details depend on the class the method KL is applied to.
}
\seealso{
\code{\link{KL.GBN}}, \code{\link{KL.CI}}, \code{\link{Fro.CI}}, \code{\link{Fro.GBN}}
}
