% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathmarks.R
\docType{data}
\name{mathmarks}
\alias{mathmarks}
\title{Math Marks Data}
\format{
A dataframe with 88 observations on the following 5 variables
\itemize{
   \item \bold{mechanics}: mark out of 100 for mechanics
   \item \bold{vectors}: mark out of 100 for vectors
   \item \bold{algebra}: mark out of 100 for algebra
   \item \bold{analysis}: mark out of 100 for analysis
   \item \bold{statistics}: mark out of 100 for statistics
}
}
\source{
Mardia, K. V., Kent, J. T. and Bibby, J. M. (1979) Multivariate Analysis. London: Academic Press.
}
\usage{
data(mathmarks)
}
\description{
Marks out of 100 for 88 students taking examinations in mechanics (C), vectors (C), algebra (0), analysis (O) and statistics (O), where C indicates closed and O indicates open book examination.
}
\keyword{datasets}
