% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pres_cov.R
\name{model_pres_cov}
\alias{model_pres_cov}
\title{Model-Preserving covariation}
\usage{
model_pres_cov(ci, type, entry, delta)
}
\arguments{
\item{ci}{object of class \code{CI}.}

\item{type}{character string. Type of model-preserving covariation: either \code{"total"}, \code{"partial"}, \code{row} or \code{column}.}

\item{entry}{a vector of length two specifying the entry of the covariance matrix to vary.}

\item{delta}{multiplicative variation coefficient for the entry of the covariance matrix given in \code{entry}.}
}
\description{
Model-Preserving covariation for objects of class \code{CI}.
}
\details{
Need to write this description, write examples and explain the return
}
\references{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\seealso{
Other gbn sensitivity: 
\code{\link{bn2}},
\code{\link{cachexia}},
\code{\link{covariance_var}()},
\code{\link{covariation_matrix}},
\code{\link{mean_var}()},
\code{\link{synthetic_cbn}}
}
\concept{gbn sensitivity}
