% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_monitors.R
\name{node}
\alias{node}
\alias{marg.node.monitor}
\alias{marg.node.monitor.graph}
\alias{cond.node.monitor}
\alias{cond.node.monitor.graph}
\alias{node.monitor}
\title{Node Monitors}
\usage{
marg.node.monitor(dag, df, node.name)

marg.node.monitor.graph(dag, df)

cond.node.monitor(dag, df, node.name)

cond.node.monitor.graph(dag, df)

node.monitor(dag, df)
}
\arguments{
\item{dag}{bnlearn object}

\item{df}{dataset}

\item{node.name}{which node}
}
\description{
Some text here
}
