% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_monitor.R
\name{node_monitor}
\alias{node_monitor}
\title{Node Monitor}
\usage{
node_monitor(dag, df, plot = TRUE)
}
\arguments{
\item{dag}{bnlearn object}

\item{df}{dataset}

\item{plot}{Boolean}
}
\description{
Some text here
}
