% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{plot.fro}
\alias{plot.fro}
\title{Plot of Frobenius}
\usage{
\method{plot}{fro}(x, ...)
}
\arguments{
\item{x}{The output of \code{Fro}}

\item{...}{additional inputs}
}
\description{
Plot of Frobenius
}
