% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_monitor.R
\name{plot.global_monitor}
\alias{plot.global_monitor}
\title{Plot for global monitors}
\usage{
\method{plot}{global_monitor}(x, ...)
}
\arguments{
\item{x}{The output of global_monitor}

\item{...}{additional inputs}
}
\description{
Plot for global monitors
}
