% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential_obs.R
\name{plot.influential_obs}
\alias{plot.influential_obs}
\title{Plot for influential observations}
\usage{
\method{plot}{influential_obs}(x, ...)
}
\arguments{
\item{x}{The output of influential_obs}

\item{...}{additional inputs}
}
\description{
Plot for influential observations
}
