% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jeffreys.R
\name{plot.jeffreys}
\alias{plot.jeffreys}
\title{Plot of  Jeffreys distance}
\usage{
\method{plot}{jeffreys}(x, ...)
}
\arguments{
\item{x}{The output of \code{Jeffreys}}

\item{...}{additional inputs}
}
\description{
Plot of  Jeffreys distance
}
