% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_monitor.R
\name{plot.node_monitor}
\alias{plot.node_monitor}
\title{Plot for node monitors}
\usage{
\method{plot}{node_monitor}(x, which, ...)
}
\arguments{
\item{x}{The output of node_monitor.}

\item{which}{select the monitor to plot, either "marginal" or "conditional".}

\item{...}{additional inputs}
}
\description{
Plot for node monitors
}
