% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot.sensitivity}
\alias{plot.sensitivity}
\title{Plot of sensitivity}
\usage{
\method{plot}{sensitivity}(x, ...)
}
\arguments{
\item{x}{The output of sensitivity}

\item{...}{additional inputs}
}
\description{
Plot of sensitivity
}
