% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_node_monitor.R
\name{plot.seq_cond_monitor}
\alias{plot.seq_cond_monitor}
\title{Plot for sequential conditional monitors}
\usage{
\method{plot}{seq_cond_monitor}(x, ...)
}
\arguments{
\item{x}{The output of seq_cond_monitor}

\item{...}{additional inputs}
}
\description{
Plot for sequential conditional monitors
}
