% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_node_monitor.R
\name{plot.seq_marg_monitor}
\alias{plot.seq_marg_monitor}
\title{Plot for sequential marginal monitors}
\usage{
\method{plot}{seq_marg_monitor}(x, ...)
}
\arguments{
\item{x}{The output of seq_marg_monitor}

\item{...}{additional inputs}
}
\description{
Plot for sequential marginal monitors
}
