% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_child_monitor.R
\name{plot.seq_pa_ch_monitor}
\alias{plot.seq_pa_ch_monitor}
\title{Plot for sequential marginal monitors}
\usage{
\method{plot}{seq_pa_ch_monitor}(x, ...)
}
\arguments{
\item{x}{The output of seq_pa_ch_monitor}

\item{...}{additional inputs}
}
\description{
Plot for sequential marginal monitors
}
