% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_distance.R
\name{print.CD}
\alias{print.CD}
\title{Print of  CD distance}
\usage{
\method{print}{CD}(x, ...)
}
\arguments{
\item{x}{The output of CD}

\item{...}{additional inputs}
}
\description{
Print of  CD distance
}
