% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{print.fro}
\alias{print.fro}
\title{Print of Frobenius norm}
\usage{
\method{print}{fro}(x, ...)
}
\arguments{
\item{x}{The output of \code{Fro}}

\item{...}{additional inputs}
}
\description{
Print of Frobenius norm
}
