% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_monitor.R
\name{print.global_monitor}
\alias{print.global_monitor}
\title{Print of global monitor}
\usage{
\method{print}{global_monitor}(x, ...)
}
\arguments{
\item{x}{The output of global_monitor}

\item{...}{additional inputs}
}
\description{
Print of global monitor
}
