% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jeffreys.R
\name{print.jeffreys}
\alias{print.jeffreys}
\title{Print of Jeffreys distance}
\usage{
\method{print}{jeffreys}(x, ...)
}
\arguments{
\item{x}{The output of Jeffreys}

\item{...}{additional inputs}
}
\description{
Print of Jeffreys distance
}
