% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_monitor.R
\name{print.node_monitor}
\alias{print.node_monitor}
\title{Print of node monitor}
\usage{
\method{print}{node_monitor}(x, ...)
}
\arguments{
\item{x}{The output of node_monitor}

\item{...}{additional inputs}
}
\description{
Print of node monitor
}
