% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.sensitivity}
\alias{print.kl}
\alias{print.CD}
\alias{print.fro}
\alias{print.global_monitor}
\alias{print.jeffreys}
\alias{print.node_monitor}
\alias{print.seq_cond_monitor}
\alias{print.seq_pa_ch_monitor}
\alias{print.seq_marg_monitor}
\title{Printing methods}
\usage{
\method{print}{sensitivity}(x, ...)

\method{print}{kl}(x, ...)

\method{print}{CD}(x, ...)

\method{print}{fro}(x, ...)

\method{print}{global_monitor}(x, ...)

\method{print}{jeffreys}(x, ...)

\method{print}{node_monitor}(x, ...)

\method{print}{seq_cond_monitor}(x, ...)

\method{print}{seq_pa_ch_monitor}(x, ...)

\method{print}{seq_marg_monitor}(x, ...)
}
\arguments{
\item{x}{an appropriate object}

\item{...}{for compatibility}
}
\value{
Printing specific to the object it is applied to.
}
\description{
Printing methods for outputs of \code{bnmonitor} functions
}
