% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{print.sensitivity}
\alias{print.sensitivity}
\title{Print of sensitivity function}
\usage{
\method{print}{sensitivity}(x, ...)
}
\arguments{
\item{x}{The output of sensitivity}

\item{...}{additional inputs}
}
\description{
Print of sensitivity function
}
