% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_node_monitor.R
\name{print.seq_marg_monitor}
\alias{print.seq_marg_monitor}
\title{Print of sequential marginal monitor}
\usage{
\method{print}{seq_marg_monitor}(x, ...)
}
\arguments{
\item{x}{The output of seq_marg_monitor}

\item{...}{additional inputs}
}
\description{
Print of sequential marginal monitor
}
