% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_child_monitor.R
\name{print.seq_pa_ch_monitor}
\alias{print.seq_pa_ch_monitor}
\title{Print of  sequential parent child monitor}
\usage{
\method{print}{seq_pa_ch_monitor}(x, ...)
}
\arguments{
\item{x}{The output of seq_pa_ch_monitor}

\item{...}{additional inputs}
}
\description{
Print of  sequential parent child monitor
}
