% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_check.R
\name{psd_check.CI}
\alias{psd_check.CI}
\title{Check for positive semidefiniteness after a perturbation for \code{CI}}
\usage{
\method{psd_check}{CI}(x, type, entry, delta, ...)
}
\arguments{
\item{x}{object of class \code{CI}.}

\item{type}{character string. Type of model-preserving covariation: either \code{"total"}, \code{"partial"}, \code{row}, \code{column} or \code{all}. If \code{all} the check is performed for every type of covariation matrix.}

\item{entry}{a vector of length 2 indicating the entry of the covariance matrix to vary.}

\item{delta}{numeric vector, including the variation parameters that act multiplicatively.}

\item{...}{additional arguments for compatibility.}
}
\value{
A dataframe including the variations performed and the check for positive semidefinitess.
}
\description{
\code{psd_check} returns a boolean to determine if the covariance matrix after a perturbation is positive semidefinite.
}
\details{
Let \eqn{\Sigma} be the covariance matrix of a Gaussian Bayesian network, \eqn{\Delta} be a perturbation matrix acting multiplicatively and \eqn{\tilde\Delta} a model-preserving covariation. The perturbed covariance matrix \eqn{\tilde\Delta\circ\Delta\circ\Sigma} is positive semidefinite if
\deqn{\rho(\tilde\Delta\circ\Delta\circ\Sigma-\Sigma)\leq \lambda_{\min}(\Sigma)}
where \eqn{\lambda_{\min}} is the smallest eigenvalue, \eqn{\rho} is the spectral radius and \eqn{\circ} is the Schur or elementwise product.
}
\examples{
psd_check(synthetic_ci,"partial",c(2,4),0.95)
psd_check(synthetic_ci,"all",c(2,3),seq(0.9,1.1,0.01))
}
\references{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\seealso{
\code{\link{psd_check.GBN}}
}
