% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_check.R
\name{psd_check.GBN}
\alias{psd_check.GBN}
\title{Check for positive semidefiniteness after a perturbation for \code{GBN}}
\usage{
\method{psd_check}{GBN}(x, entry, delta, ...)
}
\arguments{
\item{x}{object of class \code{GBN}.}

\item{entry}{a vector of length 2 indicating the entry of the covariance matrix to vary.}

\item{delta}{numeric vector, including the variation parameters that act additively.}

\item{...}{additional arguments for compatibility.}
}
\value{
A dataframe including the variations performed and the check for positive semidefinitess.
}
\description{
\code{psd_check} returns a boolean to determine if the covariance matrix after a perturbation is positive semidefinite.
}
\details{
Let \eqn{\Sigma} be the covariance matrix of a Gaussian Bayesian network and let \eqn{D} be a perturbation matrix acting additively. The perturbed covariance matrix \eqn{\Sigma+D} is positive semidefinite if
\deqn{\rho(D)\leq \lambda_{\min}(\Sigma)}
where \eqn{\lambda_{\min}} is the smallest eigenvalue end \eqn{\rho} is the spectral radius.
}
\examples{
psd_check(synthetic_gbn,c(2,4),-3)
psd_check(synthetic_gbn,c(2,3),seq(-1,1,0.1))

}
\references{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\seealso{
\code{\link{psd_check.CI}}
}
