% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_check.R
\name{psd_check}
\alias{psd_check}
\title{Check for positive semidefiniteness after a perturbation}
\usage{
psd_check(x, ...)
}
\arguments{
\item{x}{object of class \code{GBN} or \code{CI}.}

\item{...}{parameters specific to the class used.}
}
\description{
\code{psd_check} returns a boolean to determine if the covariance matrix after a perturbation is positive semidefinite.
}
\details{
The details depend on the class the method \code{KL} is applied to.
}
\seealso{
\code{\link{psd_check.GBN}}, \code{\link{psd_check.CI}}
}
