% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensquery.R
\name{sensquery}
\alias{sensquery}
\title{Sensitivity of probability query}
\usage{
sensquery(
  bnfit,
  interest_node,
  interest_node_value,
  new_value,
  evidence_nodes = NULL,
  evidence_states = NULL
)
}
\arguments{
\item{bnfit}{object of class bn.fit}

\item{interest_node}{character string. Node of the probability query of interest.}

\item{interest_node_value}{character string. Level of \code{interest_node}}

\item{new_value}{numeric value between 0 and 1. Value of the probability of interest.}

\item{evidence_nodes}{character string. Evidence nodes. Set by default to \code{NULL}.}

\item{evidence_states}{character string. Levels of \code{evidence_nodes}. If \code{NULL} no evidence is considerated. If \code{evidence_nodes="NULL"}, \code{evidence_states} should be set to \code{NULL}. Set by default to \code{NULL}.}
}
\description{
\code{sensquery} returns for a given probability of interest the parameters changes that make it possible and the resulting CD distances.
}
\details{
The Bayesian network should be expressed as a bn.fit object.
The name of the node of the probability of interest, its level and the new value should be specified. Evidence could be also indicated.
The probability of interest is specified as follows:

 P ( \code{interest_node} = \code{interest_node_value} | \code{evidence_nodes} = \code{evidence_states}  ) = \code{new_value}

By default, the  proportional covariation scheme is used for computing the CD distance.
}
\seealso{
Other sensitivity: 
\code{\link{sensitivity}()}
}
\concept{sensitivity}
