% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_node_monitor.R
\name{seq_node_monitor}
\alias{seq_node_monitor}
\alias{seq_marg_monitor}
\alias{seq_cond_monitor}
\title{Sequential Node Monitors}
\usage{
seq_marg_monitor(dag, df, node.name, plot = TRUE)

seq_cond_monitor(dag, df, node.name, plot = TRUE)
}
\arguments{
\item{dag}{bnlearn object}

\item{df}{dataset}

\item{node.name}{which node}

\item{plot}{Boolean}
}
\description{
Some text here
}
