% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential.R
\name{sequential}
\alias{sequential}
\alias{sqntl.pa.ch.monitor}
\alias{pa.ch.graph}
\title{the sequential parent child node monitor for Bayesian networks}
\usage{
sqntl.pa.ch.monitor(dframe, dag, node.name, pa.names, pa.val, alpha)

pa.ch.graph(dframe, dag, node.name, pa.names, pa.val, alpha)
}
\arguments{
\item{dframe}{base R style data frame}

\item{dag}{bnlearn object with DAG structure, not a fitte dobject}

\item{node.name}{name of child node}

\item{pa.names}{names of parents in the network}

\item{pa.val}{names of values to test}

\item{alpha}{prior}
}
\description{
the sequential parent child node monitor for Bayesian networks
}
