% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_subsidiary.R
\name{standardize}
\alias{standardize}
\title{dependent functions for the node monitors}
\usage{
standardize(vec, j)
}
\arguments{
\item{vec}{}
}
\description{
dependent functions for the node monitors
}
