% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synt_gbn.R
\docType{data}
\name{synt_gbn}
\alias{synt_gbn}
\alias{synthetic_gbn}
\alias{synthetic_ci}
\title{A synthetic continuous Bayesian network}
\format{The Bayesian network \code{bnsens_example} comprehends the following nodes:
\itemize{
  \item \bold{y1}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y2}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y3}:  three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
}}
\source{
Manuele Leonelli, Eva Riccomagno (2018). "A geometric characterisation of sensitivity analysis in monomial models". \url{https://arxiv.org/abs/1901.02058}
}
\usage{
synthetic_gbn

synthetic_ci
}
\description{
A synthetic continuous Bayesian network
}
\seealso{
Other datasets: 
\code{\link{fire_alarm}},
\code{\link{synthetic_bn}},
\code{\link{travel}}
}
\concept{datasets}
\keyword{datasets}
