% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bn.R
\docType{data}
\name{synthetic_bn}
\alias{synthetic_bn}
\title{A synthetic Bayesian network}
\format{
The Bayesian network \code{bnsens_example} comprehends the following nodes:
\itemize{
  \item \bold{y1}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y2}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y3}:  three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
}
}
\source{
Leonelli, M., & Riccomagno, E. (2022). A geometric characterization of sensitivity analysis in monomial models. International Journal of Approximate Reasoning, 151, 64-84.
}
\usage{
synthetic_bn
}
\description{
\code{synthetic_bn} is a \code{bn.fit} object for a simple Bayesian network involving three variables.
}
\keyword{datasets}
