% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cbn.R
\docType{data}
\name{synthetic_cbn}
\alias{synthetic_cbn}
\alias{synthetic_gbn}
\alias{synthetic_ci}
\title{A synthetic continuous Bayesian network}
\format{A continuous Bayesian networks over four variables ("y1", "y2", "y3", "y4"), embedding the statement "y1" independent of "y3" given "y2".
    The Bayesian network is available both as an object of class \code{GBN} and as an object of class \code{CI}.}
\source{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\usage{
synthetic_gbn

synthetic_ci
}
\description{
A synthetic continuous Bayesian network
}
\seealso{
Other datasets: 
\code{\link{cachexia}},
\code{\link{fire_alarm}},
\code{\link{synthetic_bn}},
\code{\link{travel}}

Other gbn sensitivity: 
\code{\link{bn2}},
\code{\link{cachexia}},
\code{\link{covariance_var}()},
\code{\link{covariation_matrix}},
\code{\link{mean_var}()},
\code{\link{model_pres_cov}()}
}
\concept{datasets}
\concept{gbn sensitivity}
\keyword{datasets}
