% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_covar.R
\name{uniform_covar}
\alias{uniform_covar}
\title{Uniform covariation scheme}
\usage{
uniform_covar(bnfit, node, value_node, value_parents, new_value)
}
\arguments{
\item{bnfit}{object of class bn.fit}

\item{node}{character string. Node of which the conditional probability distribution is being changed.}

\item{value_node}{character string. Level of \code{node}.}

\item{value_parents}{character string. Levels of \code{node}'s parents. The levels should be defined according to the order of the parents in \code{bnfit[[node]][["parents"]]}. If \code{node} has no parents, then should be set to \code{NULL}.}

\item{new_value}{numeric value between 0 and 1. Value to which the parameter should be updated.}
}
\description{
\code{uniform_covar} returns an updated Bayesian network using uniform covariation scheme.
}
\details{
The Bayesian network on which parameter variation is being conducted should be expressed as a bn.fit object.
The name of the node to be varied, its level and its parent's level should be specified.
The parameter variation specified by the function is:

P ( \code{node} = \code{value_node} | parents = \code{value_parents} ) = \code{new_value}
}
\concept{covariation schemes}
