% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachexia.R
\docType{data}
\name{cachexia}
\alias{cachexia}
\alias{cachexia_gbn}
\alias{cachexia_ci}
\alias{control_gbn}
\alias{control_ci}
\title{Bayesian networks for a cachexia study}
\format{Continuous Bayesian networks over six metabolomics: Adipate (A), Betaine (B), Fumarate (F), Glucose (GC), Glutamine (GM) and Valine (V).
    The networks \code{cachexia_gbn} and \code{cachexia_ci} are for people suffering of cachexia and of class \code{GBN} and \code{CI} respectively.
    The networks \code{control_gbn} and \code{control_ci} are for people not suffering of cachexia and of class \code{GBN} and \code{CI} respectively.}
\source{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\usage{
cachexia_gbn

cachexia_ci

control_gbn

control_ci
}
\description{
Continuous Bayesian networks comparing the dependence of metabolomics for people who suffer and do not suffer of Cachexia
}
\seealso{
Other datasets: 
\code{\link{fire_alarm}},
\code{\link{synthetic_bn}},
\code{\link{synthetic_cbn}},
\code{\link{travel}}

Other gbn sensitivity: 
\code{\link{bn2}},
\code{\link{covariance_var}()},
\code{\link{covariation_matrix}},
\code{\link{mean_var}()},
\code{\link{model_pres_cov}()},
\code{\link{synthetic_cbn}}
}
\concept{datasets}
\concept{gbn sensitivity}
\keyword{datasets}
