% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachexia.R
\docType{data}
\name{cachexia}
\alias{cachexia}
\alias{cachexia_gbn}
\alias{cachexia_ci}
\alias{control_gbn}
\alias{control_ci}
\alias{cachexia_data}
\title{Bayesian networks for a cachexia study}
\format{
Continuous Bayesian networks over six metabolomics: Adipate (A), Betaine (B), Fumarate (F), Glucose (GC), Glutamine (GM) and Valine (V).
    The networks \code{cachexia_gbn} and \code{cachexia_ci} are for people suffering of cachexia and of class \code{GBN} and \code{CI} respectively.
    The networks \code{control_gbn} and \code{control_ci} are for people not suffering of cachexia and of class \code{GBN} and \code{CI} respectively.
    The original dataset is stored in \code{cachexia_data}.

An object of class \code{GBN} of length 3.

An object of class \code{CI} of length 4.

An object of class \code{GBN} of length 3.

An object of class \code{CI} of length 4.

An object of class \code{data.table} (inherits from \code{data.frame}) with 77 rows and 7 columns.
}
\usage{
cachexia_gbn

cachexia_ci

control_gbn

control_ci

cachexia_data
}
\description{
Continuous Bayesian networks comparing the dependence of metabolomics for people who suffer and do not suffer of Cachexia
}
\references{
C. Görgen & M. Leonelli (2020), Model-preserving sensitivity analysis for families of Gaussian distributions.  Journal of Machine Learning Research, 21: 1-32.
}
\keyword{datasets}
