% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariation_matrix.R
\name{covariation_matrix}
\alias{covariation_matrix}
\alias{total_covar_matrix}
\alias{col_covar_matrix}
\alias{partial_covar_matrix}
\alias{row_covar_matrix}
\title{Covariation matrices}
\usage{
total_covar_matrix(ci, entry, delta)

col_covar_matrix(ci, entry, delta)

partial_covar_matrix(ci, entry, delta)

row_covar_matrix(ci, entry, delta)
}
\arguments{
\item{ci}{object of class \code{CI}.}

\item{entry}{a vector of length two specifying the entry of the covariance matrix to vary.}

\item{delta}{multiplicative variation coefficient for the entry of the covariance matrix given in \code{entry}.}
}
\description{
Construction of model-preserving matrices for objects of class \code{CI}.
}
\references{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
\seealso{
\code{\link{model_pres_cov}}
}
