% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Pima Indian Diabetes Data}
\format{
A dataframe with 392 observations on the following 9 binary variables:
\itemize{
   \item \bold{PREG}: number of times pregnant (low/high)
   \item \bold{GLUC}: plasma glucose concentration (low/high)
   \item \bold{PRES}: diastolic blood pressure (low/high)
   \item \bold{TRIC}: triceps skin fold thickness (low/high)
   \item \bold{INS}: 2-hour serum insulin (low/high)
   \item \bold{MASS}: body mass index (low/high)
   \item \bold{PED}: diabetes pedigree function (low/high)
   \item \bold{AGE}: age (low/high)
   \item \bold{DIAB}: test for diabetes (neg/pos)

}
}
\source{
These data have been taken from the UCI Repository Of Machine Learning Databases. We chose this dataset because it best showcases the function of our monitors. However, we acknowledge that this data is used here without the consent of or compensation for the original Akimel O'odham participants.
}
\description{
Discretized version of the widely-used Pima Indians Diabetes Database
}
