% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{KL.CI}
\alias{KL.CI}
\title{KL Divergence for \code{CI}}
\usage{
\method{KL}{CI}(x, type, entry, delta, ...)
}
\arguments{
\item{x}{object of class \code{CI}.}

\item{type}{character string: either \code{mean} or \code{covariance} for variations of the mean vector and covariance matrix respectively.}

\item{entry}{a vector of length 2 indicating the entry of the covariance matrix to vary.}

\item{delta}{variation parameter that acts multiplicatively.}

\item{...}{additional parameters to be added to the plot.}
}
\description{
\code{KL.CI} returns the Kullback-Leibler (KL) divergence between a Gaussian Bayesian network and its update after a model-preserving parameter variation.
}
\references{
Goergen, C., & Leonelli, M. (2018). Model-preserving sensitivity analysis for families of Gaussian distributions. arXiv preprint arXiv:1809.10794.
}
