% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_monitors.R
\name{marg.node.monitor}
\alias{marg.node.monitor}
\title{Node Monitors}
\usage{
marg.node.monitor(dag, df)
}
\arguments{
\item{dag}{bnlearn object}

\item{df}{dataset}
}
\description{
Node Monitors
}
