% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pach_monitor.R
\name{seq.pa.ch.monitor}
\alias{seq.pa.ch.monitor}
\title{the sequential parent child node monitor for Bayesian networks}
\usage{
\method{seq}{pa.ch.monitor}(df, dag, node.name, pa.names, pa.val,
  which.val)
}
\arguments{
\item{df}{base R style data frame}

\item{dag}{bnlearn object with DAG structure, not a fitte dobject}

\item{node.name}{name of child node}

\item{pa.names}{names of parents in the network}

\item{pa.val}{names of values to test}

\item{which.val}{which level of the value to test}
}
\description{
the sequential parent child node monitor for Bayesian networks
}
