% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bn.R
\docType{data}
\name{synthetic_bn}
\alias{synthetic_bn}
\title{A synthetic Bayesian network}
\format{
The Bayesian network \code{bnsens_example} comprehends the following nodes:
\itemize{
  \item \bold{y1}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y2}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y3}:  three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
}
}
\source{
Manuele Leonelli, Eva Riccomagno (2018). "A geometric characterisation of sensitivity analysis in monomial models". \url{https://arxiv.org/abs/1901.02058}
}
\usage{
synthetic_bn
}
\description{
\code{synthetic_bn} is a \code{bn.fit} object for a simple Bayesian network involving three variables.
}
\keyword{datasets}
