% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cbn.R
\docType{data}
\name{synthetic_cbn}
\alias{synthetic_cbn}
\alias{synthetic_gbn}
\alias{synthetic_ci}
\title{A synthetic continuous Bayesian network}
\format{
A continuous Bayesian networks over four variables ("y1", "y2", "y3", "y4"), embedding the statement "y1" independent of "y3" given "y2".
    The Bayesian network is available both as an object of class \code{GBN} and as an object of class \code{CI}.

An object of class \code{GBN} of length 3.

An object of class \code{CI} of length 4.
}
\source{
C. Görgen & M. Leonelli (2020), Model-preserving sensitivity analysis for families of Gaussian distributions.  Journal of Machine Learning Research, 21: 1-32.
}
\usage{
synthetic_gbn

synthetic_ci
}
\description{
A synthetic continuous Bayesian network
}
\keyword{datasets}
